+++ entry/tests/test_model.py
@@ -2242,6 +2242,7 @@ class ModelTest(AironeTestCase):
 
         resp = Entry.search_entries(user, [entity.id], [{'name': 'foo', 'keyword': 'ref'}], limit=5)
         self.assertEqual(resp['ret_count'], 10)
+        self.assertEqual(len(resp['ret_values']), 5)
 
     def test_search_entities_have_individual_attrs(self):
         user = User.objects.create(username='hoge')